pivot costs

+++ server/api_server/roadmap/models.py
@@ -32,7 +32,7 @@ class Worklogs(models.Model):
 class Worker(models.Model):
     name = models.CharField(max_length=200)
     is_active = models.BooleanField(default=False)
-    cost = models.IntegerField()
+    cost = models.FloatField()
 
 
 class RemoteLink(models.Model):
